refactor: user-service accept-pair return

+++ user-service/src/application/user/command/accept-pair/accept-pair.command-handler.ts
@@ -1,9 +1,11 @@
-import { Inject, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Inject, NotFoundException } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { AcceptPairCommand } from './accept-pair.command';
 import { UserRepository } from 'src/domain/user/repository';
 import { ClientProxy } from '@nestjs/microservices';
 import { SERVICE, ChatServiceEvent } from 'src/infrastructure/rabbitmq/service';
+import { ERROR } from 'src/infrastructure/user/common/constant';
+import { UserAggregate } from 'src/domain/user';
 
 @CommandHandler(AcceptPairCommand)
 export class AcceptPairCommandHandler
@@ -14,7 +16,7 @@ export class AcceptPairCommandHandler
     @Inject(SERVICE.CHAT) private readonly chatClient: ClientProxy,
   ) {}
 
-  async execute(command: AcceptPairCommand): Promise<string> {
+  async execute(command: AcceptPairCommand): Promise<UserAggregate> {
     const { userId, pairId } = command;
 
     const pair = await this.repository.findPair(pairId, userId);
@@ -22,10 +24,18 @@ export class AcceptPairCommandHandler
       throw new NotFoundException();
     }
 
+    const place = await this.repository.findPlace(userId);
+
+    if (!place || !pair.place) {
+      throw new BadRequestException(ERROR.NULL_PLACE);
+    }
+
+    pair.setDistanceBetweenPlaces(place);
+
     this.chatClient.emit(ChatServiceEvent.CreateChat, [userId, pairId]);
 
     await this.repository.deletePair(pair.id, userId);
 
-    return pair.id;
+    return pair;
   }
 }
